Via the configuration, one can now choose whether datatypes are restricted to core datatypes or not. If only core datatypes are required, primitive types are used where possible, otherwise BigInteger etc.
git-svn-id: svn://edison.comlab.ox.ac.uk/krr/2008/HermiT/trunk@188 6b6e81c3-f1c2-4030-a0a9-a0569132be5c

+++ src/org/semanticweb/HermiT/model/dataranges/CanonicalDataRange.java
@@ -119,7 +119,7 @@ public interface CanonicalDataRange {
      * @return true if the interpretation of this data range allows for at least 
      *         n values and false otherwise.  
      */
-    public boolean hasMinCardinality(int n);
+    public boolean hasMinCardinality(BigInteger n);
     
     /**
      * Compute the set of String representations of the allowed values for this
